# src/main/resources/latest-spark-swagger-output.yaml
...
swagger: "2.0"
info:
  description: "OpenTripPlanner Middleware API"
  title: "OTP Middleware"
  termsOfService: ""
  contact:
    name: "IBI Group"
    url: "https://github.com/ibi-group"
    email: "transitdatasupport@ibigroup.com"
  license:
    name: "MIT License"
    url: "https://opensource.org/licenses/MIT"
host: "localhost:4567"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /api/secure/monitoredtrip/checkitinerary:
    post:
      tags:
      - "api/secure/monitoredtrip"
      description: "Returns the itinerary existence check results for a monitored\
        \ trip."
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body object description"
        required: true
        schema:
          $ref: "#/definitions/MonitoredTrip"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ItineraryExistence"
          responseSchema:
            $ref: "#/definitions/ItineraryExistence"
  /api/secure/monitoredtrip:
    get:
      tags:
      - "api/secure/monitoredtrip"
      description: "Gets a paginated list of all 'MonitoredTrip' entities."
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "If specified, the maximum number of items to return."
        required: false
        type: "string"
        default: "10"
      - name: "offset"
        in: "query"
        description: "If specified, the number of records to skip/offset."
        required: false
        type: "string"
        default: "0"
      - name: "userId"
        in: "query"
        description: "If specified, the required user id."
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ResponseList"
          responseSchema:
            $ref: "#/definitions/ResponseList"
    post:
      tags:
      - "api/secure/monitoredtrip"
      description: "Creates a 'MonitoredTrip' entity."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body object description"
        required: true
        schema:
          $ref: "#/definitions/MonitoredTrip"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/MonitoredTrip"
          responseSchema:
            $ref: "#/definitions/MonitoredTrip"
...
  /api/secure/user:
    get:
      tags:
      - "api/secure/user"
      description: "Gets a paginated list of all 'OtpUser' entities."
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "If specified, the maximum number of items to return."
        required: false
        type: "string"
        default: "10"
      - name: "offset"
        in: "query"
        description: "If specified, the number of records to skip/offset."
        required: false
        type: "string"
        default: "0"
      - name: "userId"
        in: "query"
        description: "If specified, the required user id."
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ResponseList"
          responseSchema:
            $ref: "#/definitions/ResponseList"
    post:
      tags:
      - "api/secure/user"
      description: "Creates a 'OtpUser' entity."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Body object description"
        required: true
        schema:
          $ref: "#/definitions/OtpUser"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/OtpUser"
          responseSchema:
            $ref: "#/definitions/OtpUser"
...
definitions:
...
  ResponseList:
    type: "object"
    properties:
      clazz:
        type: "string"
      data:
        type: "array"
        items:
          $ref: "#/definitions/Object"
      offset:
        type: "integer"
        format: "int32"
      limit:
        type: "integer"
        format: "int32"
      total:
        type: "integer"
        format: "int64"
      timestamp:
        type: "string"
        format: "date"
...
  TripMonitorNotification:
    type: "object"
    properties:
      type:
        type: "string"
        enum:
        - "DEPARTURE_DELAY"
        - "ARRIVAL_DELAY"
        - "ITINERARY_CHANGED"
        - "ALERT_FOUND"
        - "ITINERARY_NOT_FOUND"
      body:
        type: "string"
  EncodedPolyline:
    type: "object"
    properties:
      points:
        type: "string"
      levels:
        type: "string"
      length:
        type: "integer"
        format: "int32"
  FareDetails:
    type: "object"
    properties:
      regular:
        type: "array"
        items:
          $ref: "#/definitions/FareComponent"
      student:
        type: "array"
        items:
          $ref: "#/definitions/FareComponent"
      senior:
        type: "array"
        items:
          $ref: "#/definitions/FareComponent"
      tram:
        type: "array"
        items:
          $ref: "#/definitions/FareComponent"
      special:
        type: "array"
        items:
          $ref: "#/definitions/FareComponent"
      youth:
        type: "array"
        items:
          $ref: "#/definitions/FareComponent"
  Step:
    type: "object"
    properties:
      distance:
        type: "number"
        format: "double"
      relativeDirection:
        type: "string"
      streetName:
        type: "string"
      absoluteDirection:
        type: "string"
      stayOn:
        type: "boolean"
      area:
        type: "boolean"
      bogusName:
        type: "boolean"
      lon:
        type: "number"
        format: "double"
      lat:
        type: "number"
        format: "double"
  ItineraryExistenceResult:
    type: "object"
    properties:
      invalidDates:
        type: "array"
        items:
          type: "string"
      validDates:
        type: "array"
        items:
          type: "string"
      itineraries:
        type: "array"
        items:
          $ref: "#/definitions/Itinerary"
  Fare:
    type: "object"
    properties:
      regular:
        $ref: "#/definitions/Price"
      student:
        $ref: "#/definitions/Price"
      senior:
        $ref: "#/definitions/Price"
      tram:
        $ref: "#/definitions/Price"
      special:
        $ref: "#/definitions/Price"
      youth:
        $ref: "#/definitions/Price"
  Leg:
    type: "object"
    properties:
      startTime:
        type: "string"
        format: "date"
      endTime:
        type: "string"
        format: "date"
      departureDelay:
        type: "integer"
        format: "int32"
      arrivalDelay:
        type: "integer"
        format: "int32"
      realTime:
        type: "boolean"
      distance:
        type: "number"
        format: "double"
      pathway:
        type: "boolean"
      mode:
        type: "string"
      route:
        type: "string"
      interlineWithPreviousLeg:
        type: "boolean"
      from:
        $ref: "#/definitions/Place"
      to:
        $ref: "#/definitions/Place"
      legGeometry:
        $ref: "#/definitions/EncodedPolyline"
      rentedBike:
        type: "boolean"
      rentedCar:
        type: "boolean"
      rentedVehicle:
        type: "boolean"
      hailedCar:
        type: "boolean"
      transitLeg:
        type: "boolean"
      duration:
        type: "number"
        format: "double"
      intermediateStops:
        type: "array"
        items:
          $ref: "#/definitions/Place"
      steps:
        type: "array"
        items:
          $ref: "#/definitions/Step"
      agencyName:
        type: "string"
      agencyUrl:
        type: "string"
      routeType:
        type: "integer"
        format: "int32"
      routeId:
        type: "string"
      agencyId:
        type: "string"
      tripBlockId:
        type: "string"
      tripId:
        type: "string"
      serviceDate:
        type: "string"
      interStopGeometry:
        type: "array"
        items:
          $ref: "#/definitions/EncodedPolyline"
      routeShortName:
        type: "string"
      routeLongName:
        type: "string"
      alerts:
        type: "array"
        items:
          $ref: "#/definitions/LocalizedAlert"
      headsign:
        type: "string"
  OtpRequest:
    type: "object"
    properties:
      dateTime:
        $ref: "#/definitions/ZonedDateTime"
      requestParameters:
        $ref: "#/definitions/Map"
  Price:
    type: "object"
    properties:
      currency:
        $ref: "#/definitions/Currency"
      cents:
        type: "integer"
        format: "int32"
  ItineraryExistence:
    type: "object"
    properties:
      otpRequests:
        type: "array"
        items:
          $ref: "#/definitions/OtpRequest"
      referenceItinerary:
        $ref: "#/definitions/Itinerary"
      monday:
        $ref: "#/definitions/ItineraryExistenceResult"
      tuesday:
        $ref: "#/definitions/ItineraryExistenceResult"
      wednesday:
        $ref: "#/definitions/ItineraryExistenceResult"
      thursday:
        $ref: "#/definitions/ItineraryExistenceResult"
      friday:
        $ref: "#/definitions/ItineraryExistenceResult"
      saturday:
        $ref: "#/definitions/ItineraryExistenceResult"
      sunday:
        $ref: "#/definitions/ItineraryExistenceResult"
      message:
        type: "string"
      error:
        type: "boolean"
      tripIsArriveBy:
        type: "boolean"
      timestamp:
        type: "string"
        format: "date"
  Currency:
    type: "object"
    properties:
      symbol:
        type: "string"
      currency:
        type: "string"
      defaultFractionDigits:
        type: "integer"
        format: "int32"
      currencyCode:
        type: "string"
  JourneyState:
    type: "object"
    properties:
      baselineArrivalTimeEpochMillis:
        type: "integer"
        format: "int64"
      baselineDepartureTimeEpochMillis:
        type: "integer"
        format: "int64"
      scheduledArrivalTimeEpochMillis:
        type: "integer"
        format: "int64"
      scheduledDepartureTimeEpochMillis:
        type: "integer"
        format: "int64"
      lastCheckedEpochMillis:
        type: "integer"
        format: "int64"
      lastNotifications:
        type: "array"
        items:
          $ref: "#/definitions/TripMonitorNotification"
      lastNotificationTimeMillis:
        type: "integer"
        format: "int64"
      matchingItinerary:
        $ref: "#/definitions/Itinerary"
      targetDate:
        type: "string"
      tripStatus:
        type: "string"
        enum:
        - "NO_LONGER_POSSIBLE"
        - "NEXT_TRIP_NOT_POSSIBLE"
        - "TRIP_UPCOMING"
        - "TRIP_ACTIVE"
      hasRealtimeData:
        type: "boolean"
  Itinerary:
    type: "object"
    properties:
      duration:
        type: "integer"
        format: "int64"
      startTime:
        type: "string"
        format: "date"
      endTime:
        type: "string"
        format: "date"
      walkTime:
        type: "integer"
        format: "int64"
      transitTime:
        type: "integer"
        format: "int64"
      waitingTime:
        type: "integer"
        format: "int64"
      walkDistance:
        type: "number"
        format: "double"
      walkLimitExceeded:
        type: "boolean"
      elevationLost:
        type: "number"
        format: "double"
      elevationGained:
        type: "number"
        format: "double"
      transfers:
        type: "integer"
        format: "int32"
      fare:
        $ref: "#/definitions/FareWrapper"
      legs:
        type: "array"
        items:
          $ref: "#/definitions/Leg"
  FareComponent:
    type: "object"
    properties:
      fareId:
        type: "string"
      price:
        $ref: "#/definitions/Price"
      routes:
        type: "array"
        items:
          type: "string"
  FareWrapper:
    type: "object"
    properties:
      fare:
        $ref: "#/definitions/Fare"
      details:
        $ref: "#/definitions/FareDetails"
  Place:
    type: "object"
    properties:
      name:
        type: "string"
      lon:
        type: "number"
        format: "double"
      lat:
        type: "number"
        format: "double"
      departure:
        type: "string"
        format: "date"
      orig:
        type: "string"
      vertexType:
        type: "string"
      stopId:
        type: "string"
      arrival:
        type: "string"
        format: "date"
      stopIndex:
        type: "integer"
        format: "int32"
      stopSequence:
        type: "integer"
        format: "int32"
      stopCode:
        type: "string"
      platformCode:
        type: "string"
      zoneId:
        type: "string"
      bikeShareId:
        type: "string"
      networks:
        type: "array"
        items:
          type: "string"
      address:
        type: "string"
  LocalizedAlert:
    type: "object"
    properties:
      alertHeaderText:
        type: "string"
      alertDescriptionText:
        type: "string"
      alertUrl:
        type: "string"
      effectiveStartDate:
        type: "string"
        format: "date"
  MonitoredTrip:
    type: "object"
    properties:
      userId:
        type: "string"
      tripName:
        type: "string"
      tripTime:
        type: "string"
      from:
        $ref: "#/definitions/Place"
      to:
        $ref: "#/definitions/Place"
      arriveBy:
        type: "boolean"
      leadTimeInMinutes:
        type: "integer"
        format: "int32"
      monday:
        type: "boolean"
      tuesday:
        type: "boolean"
      wednesday:
        type: "boolean"
      thursday:
        type: "boolean"
      friday:
        type: "boolean"
      saturday:
        type: "boolean"
      sunday:
        type: "boolean"
      excludeFederalHolidays:
        type: "boolean"
      isActive:
        type: "boolean"
      snoozed:
        type: "boolean"
      queryParams:
        type: "string"
      itinerary:
        $ref: "#/definitions/Itinerary"
      notifyOnAlert:
        type: "boolean"
      departureVarianceMinutesThreshold:
        type: "integer"
        format: "int32"
      arrivalVarianceMinutesThreshold:
        type: "integer"
        format: "int32"
      notifyOnItineraryChange:
        type: "boolean"
      itineraryExistence:
        $ref: "#/definitions/ItineraryExistence"
      journeyState:
        $ref: "#/definitions/JourneyState"
...
  UserLocation:
    type: "object"
    properties:
      address:
        type: "string"
      icon:
        type: "string"
      lat:
        type: "number"
        format: "double"
      lon:
        type: "number"
        format: "double"
      name:
        type: "string"
      type:
        type: "string"
  OtpUser:
    type: "object"
    properties:
      accessibilityRoutingByDefault:
        type: "boolean"
      hasConsentedToTerms:
        type: "boolean"
      isPhoneNumberVerified:
        type: "boolean"
      notificationChannel:
        type: "string"
      phoneNumber:
        type: "string"
      savedLocations:
        type: "array"
        items:
          $ref: "#/definitions/UserLocation"
      storeTripHistory:
        type: "boolean"
      applicationId:
        type: "string"
...